requirements.txt
fastapi==0.115.2
importlib_resources==6.4.5
uvicorn==0.31.1
psycopg==3.2.3
psycopg_pool==3.2.3
pydantic_settings==2.5.2